Update to Dalamud API Level 15 and add roll tracking toggle
Fixes breaking API changes from Dalamud 15 (IClientState.LocalPlayer/LocalContentId
moved to IObjectTable/IPlayerState, IChatGui.ChatMessage delegate signature, uint
TerritoryType), bumps to v1.3.3, and adds a config option to disable the
Need/Greed roll popup for players who don't want it.

+++ src/Services/LootTableService.cs
@@ -130,7 +130,7 @@ public class LootTableService : IDisposable
     {
         try
         {
-            var territoryId = Plugin.ClientState.TerritoryType;
+            var territoryId = (ushort)Plugin.ClientState.TerritoryType;
             if (territoryId == 0)
             {
                 Plugin.Log.Debug("Not in a valid territory");
